reuse log_finished
